requirements.txt
# === Backend Core ===
fastapi==0.111.0
uvicorn[standard]==0.29.0

# === Quantum AI Core ===
pennylane==0.35.1
pennylane-lightning==0.35.1  # ⬅️ Faster backend for simulations
qiskit==1.0.2

# === Machine Learning / Transformers ===
torch==2.2.2
transformers==4.40.1
accelerate==0.29.2
scikit-learn==1.4.2

# === HuggingFace Utilities ===
huggingface_hub==0.22.2
datasets==2.19.1
tokenizers==0.19.1  # Optional but improves tokenizer performance

# === Serving + Dev Tools ===
python-dotenv==1.0.1
requests==2.31.0
httpx==0.27.0
aiofiles==23.2.1
jinja2==3.1.3
python-multipart==0.0.9  # (for form-based file uploads)

# === Logging / Debugging (Optional) ===
rich==13.7.1
loguru==0.7.2

# === Optional Dev Tools ===
ipython==8.23.0
black==24.4.0
ruff==0.3.5
isort==5.13.2
pre-commit==3.6.0
mypy==1.9.0

# === Optional for Data ===
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
plotly==5.24.0
